[FEATURE] Add setting to decide which form fields are marked as required
The new setting plugin.tx_reserve.settings.requiredFormFields is a comma
separated list that includes the field names to be required in frontend
reservation form. You can use the constant editor or the path to that
setting to add your own fields or just update the default list of required
fields.